requirements.txt
fastapi>=0.104.1
uvicorn>=0.24.0
pydantic>=2.5.2
sentence-transformers>=2.2.2
faiss-cpu
rapidfuzz>=3.5.2
symspellpy>=6.7.7
scikit-learn
joblib
pyyaml
numpy
